+++ rgb-solver/src/solver/tree_path/mod.rs
@@ -1,1 +1,3 @@
-pub mod tree;
+mod path;
+
+pub use self::path::calc_paths;
